Refactor the utilities for time windows and add unit tests, #130
