+++ wres-datamodel/src/wres/datamodel/time/TimeSeriesSlicer.java
@@ -1616,84 +1616,6 @@ public final class TimeSeriesSlicer
     }
 
     /**
-     * Subtracts any non-instantaneous desired timescale from the earliest lead duration and the earliest valid time to
-     * ensure that sufficient data is retrieved for upscaling.
-     *
-     * @param timeWindow the time window to adjust
-     * @param timeScale the timescale to use
-     * @return the adjusted time window
-     */
-
-    public static TimeWindowOuter adjustTimeWindowForTimeScale( TimeWindowOuter timeWindow,
-                                                                TimeScaleOuter timeScale )
-    {
-        TimeWindow.Builder adjusted = timeWindow.getTimeWindow()
-                                                .toBuilder();
-
-        // Earliest lead duration
-        if ( !timeWindow.getEarliestLeadDuration()
-                        .equals( TimeWindowOuter.DURATION_MIN ) )
-        {
-            Duration period = Duration.ZERO;
-
-            // Adjust the lower bound of the lead duration window by the non-instantaneous desired timescale
-            if ( Objects.nonNull( timeScale )
-                 && !timeScale.isInstantaneous() )
-            {
-                period = TimeScaleOuter.getOrInferPeriodFromTimeScale( timeScale );
-            }
-
-            Duration lowered = timeWindow.getEarliestLeadDuration()
-                                         .minus( period );
-
-            if ( Objects.nonNull( timeScale )
-                 && LOGGER.isDebugEnabled() )
-            {
-                LOGGER.debug( "Adjusting the lower lead duration of time window {} from {} to {} "
-                              + "in order to acquire data at the desired timescale of {}.",
-                              timeWindow,
-                              timeWindow.getEarliestLeadDuration(),
-                              lowered,
-                              timeScale );
-            }
-
-            adjusted.setEarliestLeadDuration( MessageFactory.getDuration( lowered ) );
-        }
-
-        // Earliest valid time
-        if ( !timeWindow.getEarliestValidTime()
-                        .equals( Instant.MIN ) )
-        {
-            Duration period = Duration.ZERO;
-
-            // Adjust the lower bound of the lead duration window by the non-instantaneous desired timescale
-            if ( Objects.nonNull( timeScale )
-                 && !timeScale.isInstantaneous() )
-            {
-                period = TimeScaleOuter.getOrInferPeriodFromTimeScale( timeScale );
-            }
-
-            Instant lowered = timeWindow.getEarliestValidTime()
-                                        .minus( period );
-
-            if ( Objects.nonNull( timeScale )
-                 && LOGGER.isDebugEnabled() )
-            {
-                LOGGER.debug( "Adjusting the lower valid datetime of time window {} from {} to {} "
-                              + "in order to acquire data at the desired timescale of {}.",
-                              timeWindow,
-                              timeWindow.getEarliestValidTime(),
-                              lowered,
-                              timeScale );
-            }
-
-            adjusted.setEarliestValidTime( MessageFactory.getTimestamp( lowered ) );
-        }
-
-        return TimeWindowOuter.of( adjusted.build() );
-    }
-
-    /**
      * Adds a prescribed offset to the valid time of each time-series in the list.
      *
      * @param <T> the time-series event value type
