+++ wres-datamodel/test/wres/datamodel/time/TimeSeriesSlicerTest.java
@@ -1461,64 +1461,6 @@ final class TimeSeriesSlicerTest
     }
 
     @Test
-    void testAdjustByTimeScalePeriodWhenTimeScaleIsInstantaneous()
-    {
-        TimeWindow timeWindow = MessageFactory.getTimeWindow( Duration.ofHours( 1 ),
-                                                              Duration.ofHours( 2 ) );
-        TimeScale timeScale = TimeScale.newBuilder()
-                                       .setPeriod( com.google.protobuf.Duration.newBuilder()
-                                                                               .setSeconds( 1 ) )
-                                       .build();
-        TimeScaleOuter timeScaleOuter = TimeScaleOuter.of( timeScale );
-        TimeWindowOuter timeWindowOuter = TimeWindowOuter.of( timeWindow );
-
-        TimeWindowOuter actual = TimeSeriesSlicer.adjustTimeWindowForTimeScale( timeWindowOuter, timeScaleOuter );
-        assertEquals( timeWindowOuter, actual );
-    }
-
-    @Test
-    void testAdjustTimeWindowEarliestLeadDurationForTimeScale()
-    {
-        TimeWindow timeWindow = MessageFactory.getTimeWindow( Duration.ofHours( 1 ),
-                                                              Duration.ofHours( 2 ) );
-        TimeScale timeScale = TimeScale.newBuilder()
-                                       .setPeriod( com.google.protobuf.Duration.newBuilder()
-                                                                               .setSeconds( 1800 ) )
-                                       .build();
-        TimeScaleOuter timeScaleOuter = TimeScaleOuter.of( timeScale );
-        TimeWindowOuter timeWindowOuter = TimeWindowOuter.of( timeWindow );
-
-        TimeWindowOuter actual = TimeSeriesSlicer.adjustTimeWindowForTimeScale( timeWindowOuter, timeScaleOuter );
-        TimeWindow expectedInner = MessageFactory.getTimeWindow( Duration.ofMinutes( 30 ),
-                                                                 Duration.ofHours( 2 ) );
-        TimeWindowOuter expected = TimeWindowOuter.of( expectedInner );
-
-        assertEquals( expected, actual );
-    }
-
-    @Test
-    void testAdjustTimeWindowEarliestValidTimeForTimeScale()
-    {
-        Instant earliest = Instant.parse( "2055-03-23T00:00:00Z" );
-        Instant latest = Instant.parse( "2055-03-24T00:00:00Z" );
-
-        TimeWindow timeWindow = MessageFactory.getTimeWindow( earliest, latest );
-        TimeScale timeScale = TimeScale.newBuilder()
-                                       .setPeriod( com.google.protobuf.Duration.newBuilder()
-                                                                               .setSeconds( 86400 ) )
-                                       .build();
-        TimeScaleOuter timeScaleOuter = TimeScaleOuter.of( timeScale );
-        TimeWindowOuter timeWindowOuter = TimeWindowOuter.of( timeWindow );
-
-        TimeWindowOuter actual = TimeSeriesSlicer.adjustTimeWindowForTimeScale( timeWindowOuter, timeScaleOuter );
-        TimeWindow expectedInner = MessageFactory.getTimeWindow( Instant.parse( "2055-03-22T00:00:00Z" ),
-                                                                 latest );
-        TimeWindowOuter expected = TimeWindowOuter.of( expectedInner );
-
-        assertEquals( expected, actual );
-    }
-
-    @Test
     void testGetTimesteps()
     {
         TimeSeries<Double> series =
